requirements.txt
aiohttp>=3.9.0
aiosignal==1.3.1
altair==5.1.1
async-timeout==4.0.3
attrs==23.1.0
audio-recorder-streamlit==0.0.8
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
frozenlist==1.4.0
gitdb==4.0.10
GitPython==3.1.35
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
langchain>=0.1.0
langchain-google-genai==0.0.5
langchain-community>=0.0.10
mdurl==0.1.2
multidict==6.0.4
numpy>=1.26.0
openai==0.28.0
packaging>=23.2,<24.0
pandas>=2.1.0
pdfminer.six>=20221105
Pillow>=10.0.0
protobuf==4.24.3
pyarrow>=14.0.1
pycparser==2.21
pydeck==0.8.0
pydub==0.25.1
Pygments==2.16.1
Pympler==1.0.1
python-dotenv==1.0.0
python-dateutil==2.8.2
python-magic>=0.4.27
pytz==2023.3.post1
pytz-deprecation-shim==0.1.0.post0
referencing==0.30.2
requests==2.31.0
rpds-py==0.10.2
six==1.16.0
smmap==5.0.0
sounddevice==0.4.6
soundfile==0.12.1
streamlit>=1.24.0
streamlit-audiorecorder==0.0.3.post1
streamlit-extras>=0.3.0
tenacity==8.2.3
toml==0.10.2
toolz==0.12.0
tzdata==2023.3
tzlocal==4.3.1
urllib3==2.0.4
openai
ffprobe
unstructured[pdf]>=0.10.0
setuptools>=69.0.0
wheel>=0.42.0
